Aligned NonbondedInteractionGroup and NonbondedAllPairs potentials (#1538)
* Refactor SingleTopology to use aligned nonbonded potentials.

* Clean-up stale code.

* Bugfixes for nonbonded interaction group.

* Update comments.

* fix nn (#1541)

* Round trip to 32-bit and back

* Add assertions for mypy

* Add tests for du_dx

* Stronger equality tests without casting trick.

* Fix stray comments.

* Update AlignedPotential with ABC, post_init, rename mins, maxes to lambda_min, lambda_max

* Update comment on linear_interpolate use of jnp.where

* Update more comments.

* WIP

* Fix Nones

* Update hashes.

* Update docstring.

* PR Feedback: incorporate potential directly into class.

* Simplify typing.

* Update hashes.

---------

+++ tests/test_single_topology_combined.py
@@ -216,4 +216,6 @@ def test_nonbonded_host_params_independent_of_lambda(
     params0 = get_nonbonded_host_params(0.0)
     for lamb in np.linspace(0.1, 1, 10):
         params = get_nonbonded_host_params(lamb)
-        np.testing.assert_array_equal(params, params0)
+        np.testing.assert_array_equal(
+            params.astype(np.float32).astype(np.float64), params0.astype(np.float32).astype(np.float64)
+        )
